+++ hrflow/tracking/__init__.py
@@ -1,13 +1,14 @@
-from hrflow.hrflow.utils import validate_response
+from ..utils import validate_response
 
 
-class Rating:
+class Tracking:
     def __init__(self, client):
         self.client = client
 
     def get(
         self,
         role,
+        actions=None,
         source_keys=None,
         source_key=None,
         profile_key=None,
@@ -36,16 +37,15 @@ class Rating:
         educations_duration_max=None,
     ):
         """
-        This endpoint allows you to retrieve the list of ratings.
+        This endpoint allows you to retrieve the list of trackings.
         Visit : https://developers.hrflow.ai/reference for more information.
 
-        Retrieve Ratings Associated with a Specific Profile or Job
+        Retrieve Trackings Associated with a Specific Profile or Job
             - To filter by a specific profile, include the parameters (source_key,
                 profile_key//profile_reference) and leave source_keys empty.
             - To filter by a specific job, include the parameters (board_key,
                 job_key//job_reference) and leave board_keys empty.
-
-        Retrieve Ratings Associated with a Specific List of Profiles or Jobs
+        Retrieve Trackings Associated with a Specific List of Profiles or Jobs
             - To filter ratings based on a list of profiles within a list of sources,
                 include the parameter source_keys and leave (source_key,
                 profile_key//profile_reference) empty.
@@ -53,20 +53,21 @@ class Rating:
                 include the parameter board_keys and leave (board_key,
                 job_key//job_reference) empty.
         """
-        args = locals()
+        args = locals()  # get all the arguments (local variables) of the function
         params = {}
 
         for arg, value in args.items():
             if value is not None and arg != "self":
                 params[arg] = value
 
-        response = self.client.get(resource_endpoint="ratings", query_params=params)
+        # Underlying resource : GET /trackings
+        response = self.client.get(resource_endpoint="trackings", query_params=params)
 
         return validate_response(response)
 
     def post(
         self,
-        score,
+        action,
         role,
         board_key,
         source_key,
@@ -79,47 +80,51 @@ class Rating:
         created_at=None,
     ):
         """
-        This endpoint allows you to rate a Profile (resp. a Job) for Job (resp. a
-        Profile)
-        as a recruiter (resp. a candidate) with a score between 0 and 1.
+        This endpoint allows you to track a Profile (resp. a Job) for Job (resp. a
+        Profile) as a recruiter
+        (resp. a candidate) with a specific action
+
         Visit : https://developers.hrflow.ai/reference for more information.
 
-        The job_key and the job_reference cannot be null at the same time in the
-        Request Parameters.
-        The same for the profile_key and profile_reference.
+        Note : The job_key and the job_reference cannot be null at the same time in
+        the Request Parameters.
+        The same for the profile_key and profile_reference .
 
         Args:
-            score:                  <float>
-                                    The score is an evaluation fit between 0 to 1 .
-                                    If you're using stars in your system please use
-                                    the following conversion:
-                                    5 stars = 1.0 , 4 stars=0.8 , 3 stars=0.6, 2
-                                    stars=0.4, 1 star= 0.2 .
+            action:                 <string>
+                                    The 'action' refers to a unique identifier for a
+                                    profile or job stage.
+                                    This can be a specific stage ID within a CRM,
+                                    ATS, or Job site.
+                                    Examples of such stages include "view," "apply,"
+                                    "hire," or any other stage relevant to your system.
+
             role:                   <string>
-                                    Role of the user rating the job role in
-                                    {recruiter, candidate, employee, manager}.
+                                    Role of the user rating the job (role: recruiter,
+                                    candidate, employee, manager).
             board_key:              <string>
                                     The key of Board attached to the given Job.
             source_key:             <string>
                                     The key of Source attached to the given Profile.
             job_key:                <string>
-                                    job identifier (key)
+                                    The Job's unique identifier.
             job_reference:          <string>
-                                    The Job reference chosen by the customer or an
+                                    The Job's reference chosen by the customer /
                                     external system.
                                     If you use the job_key you do not need to specify
                                     the job_reference and vice versa.
             profile_key:            <string>
-                                    profile identifier (key)
+                                    The Profile's unique identifier.
             profile_reference:      <string>
-                                    The Profile reference chosen by the customer or an
+                                    The Profile's reference chosen by the customer /
                                     external system.
                                     If you use the profile_key you do not need to
                                     specify the profile_reference and vice versa.
             author_email:           <string>
-                                    author email
+                                    Email of the HrFlow.ai user who rated the profile
+                                    for the job.
             comment:                <string>
-                                    comment
+                                    Comment explaining the reason behind the score.
             created_at:             <string>
                                     ISO Date of the rating.
                                     Format : yyyy-MM-dd'T'HH:mm:ss.SSSXXX — for
@@ -129,8 +134,8 @@ class Rating:
                                     application of the profile).
                                     If not provided the creation date will be now by
                                     default.
-        """
 
+        """
         args = locals()
         body = {}
 
@@ -138,6 +143,6 @@ class Rating:
             if value is not None and arg != "self":
                 body[arg] = value
 
-        # Underlying resource : POST /rating
-        response = self.client.post(resource_endpoint="rating", json=body)
+        # Underlying resource : POST /tracking
+        response = self.client.post(resource_endpoint="tracking", json=body)
         return validate_response(response)
